added auth levels to bids

+++ website/views/bid.py
@@ -132,6 +132,8 @@ def get_winner_by_auction_id(request, auction_id):
 @csrf_exempt
 @api_view(['POST'])
 def update_one_bid(request, _id):
+    if not request.user.is_staff:
+        return JsonResponse({"message": "Unauthorized"}, status=401)
     data = json.loads(request.body)
     bid = {
         "auctionId": data.get('auctionId'),
@@ -151,6 +153,8 @@ def update_one_bid(request, _id):
 @csrf_exempt
 @api_view(['DELETE'])
 def delete_one_bid(request, _id):
+    if not request.user.is_staff:
+        return JsonResponse({"message": "Unauthorized"}, status=401)
     url = f'{host}deleteOneBid/{_id}'
     response = requests.delete(url, headers=bid_get_token_middleware(request))
     data = response.json()
@@ -160,6 +164,8 @@ def delete_one_bid(request, _id):
 @csrf_exempt
 @api_view(['DELETE'])
 def delete_all_bids_by_auction_id(request, auction_id):
+    if not request.user.is_staff:
+        return JsonResponse({"message": "Unauthorized"}, status=401)
     url = f'{host}deleteAllBidsByAuctionId/{auction_id}'
     response = requests.delete(url, headers=bid_get_token_middleware(request))
     data = response.json()
